update to `newtype_rt_variable_hash!`
