--- kupyna/src/long_compress.rs
@@ -1,95 +1,0 @@
-use crate::utils::{add_constant_plus, add_constant_xor, apply_s_box, mix_columns, xor_bytes};
-
-pub(crate) const COLS: usize = 16;
-const ROUNDS: u64 = 14;
-
-type Matrix = [[u8; 8]; 16];
-
-pub(crate) fn compress(prev_vector: &mut [u64; COLS], message_block: &[u8; 128]) {
-    let mut prev_vector_u8 = [0u8; 128];
-    for (src, dst) in prev_vector.iter().zip(prev_vector_u8.chunks_exact_mut(8)) {
-        dst.copy_from_slice(&src.to_be_bytes());
-    }
-
-    let m_xor_p = xor_bytes(*message_block, prev_vector_u8);
-
-    let t_xor_mp = t_xor_l(m_xor_p);
-
-    let t_plus_m = t_plus_l(*message_block);
-
-    prev_vector_u8 = xor_bytes(xor_bytes(t_xor_mp, t_plus_m), prev_vector_u8);
-
-    for (dst, src) in prev_vector.iter_mut().zip(prev_vector_u8.chunks_exact(8)) {
-        *dst = u64::from_be_bytes(src.try_into().unwrap());
-    }
-}
-
-pub(crate) fn t_plus_l(block: [u8; 128]) -> [u8; 128] {
-    let mut state = block_to_matrix(block);
-    for nu in 0..ROUNDS {
-        state = add_constant_plus(state, nu as usize);
-        state = apply_s_box(state);
-        state = rotate_rows(state);
-        state = mix_columns(state);
-    }
-    matrix_to_block(state)
-}
-
-fn block_to_matrix(block: [u8; 128]) -> Matrix {
-    const ROWS: usize = 16;
-    const COLS: usize = 8;
-
-    let mut matrix = [[0u8; COLS]; ROWS];
-    for i in 0..ROWS {
-        for j in 0..COLS {
-            matrix[i][j] = block[i * COLS + j];
-        }
-    }
-    matrix
-}
-
-fn matrix_to_block(matrix: Matrix) -> [u8; 128] {
-    const ROWS: usize = 16;
-    const COLS: usize = 8;
-
-    let mut block = [0u8; ROWS * COLS];
-    for i in 0..ROWS {
-        for j in 0..COLS {
-            block[i * COLS + j] = matrix[i][j];
-        }
-    }
-    block
-}
-
-fn rotate_rows(mut state: Matrix) -> Matrix {
-    const ROWS: usize = 16;
-    let cols = 8;
-
-    let mut temp = [0u8; ROWS];
-    let mut shift: i32 = -1;
-    for i in 0..cols {
-        if i == cols - 1 {
-            shift = 11;
-        } else {
-            shift += 1;
-        }
-        for col in 0..ROWS {
-            temp[(col + shift as usize) % ROWS] = state[col][i];
-        }
-        for col in 0..ROWS {
-            state[col][i] = temp[col];
-        }
-    }
-    state
-}
-
-pub(crate) fn t_xor_l(block: [u8; 128]) -> [u8; 128] {
-    let mut state = block_to_matrix(block);
-    for nu in 0..ROUNDS {
-        state = add_constant_xor(state, nu as usize);
-        state = apply_s_box(state);
-        state = rotate_rows(state);
-        state = mix_columns(state);
-    }
-    matrix_to_block(state)
-}
